ex 4.21

+++ controllers/blogs.js
@@ -44,8 +44,19 @@ blogsRouter.post('/', async (request, response) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  const blog = await Blog.findById(request.params.id);
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  const userid = decodedToken.id;
+  if (!userid) {
+    return response.status(401).json({ error: 'token missing or invalid' });
+  }
+  // const user = await User.findById(userid);
+  if (blog.user.toString() === userid) {
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  } else {
+    response.status(401).json({ error: 'wrong user' });
+  }
 });
 
 blogsRouter.put('/:id', async (request, response) => {
